Update README.md
First, we split the string using space regex and created an array of strings. Then we used a map to make an entry, where the word is the key, and its frequency is the value. Later we sorted the map based on the frequency value to show the entries in descending order